+++ yggdrasil/communication/RMQComm.py
@@ -244,15 +244,20 @@ class RMQComm(CommBase.CommBase):
             kwargs['address'] = _rmq_param_sep.join([url, exchange, queue])
         return args, kwargs
 
-    def opp_comm_kwargs(self):
+    def opp_comm_kwargs(self, for_yaml=False):
         r"""Get keyword arguments to initialize communication with opposite
         comm object.
+
+        Args:
+            for_yaml (bool, optional): If True, the returned dict will only
+                contain values that can be specified in a YAML file. Defaults
+                to False.
 
         Returns:
             dict: Keyword arguments for opposite comm object.
 
         """
-        kwargs = super(RMQComm, self).opp_comm_kwargs()
+        kwargs = super(RMQComm, self).opp_comm_kwargs(for_yaml=for_yaml)
         return kwargs
 
     def bind(self):